proje/requirements.txt
mysql-connector-python
flask
flask-mysql
